[SRP] Uma classe para gerar PDF
